Avoid "Flash of Invisible Text" with`display=swap`
Google now allows you to provide a `display` parameter when importing fonts and uses that value for the `@font-face` `font-display` descriptor. This tells the browser how to display the font if it's not yet downloaded. Google now defaults to `display=swap`, which displays fonts using a fallback at first, then swaps in the requested font once it is loaded.